requirements.txt
Django
django-cors-headers
djangorestframework
graphene
graphene-django
graphql-core
graphql-relay
iso8601
promise
psycopg2
pytz
singledispatch
six
typing
requests